Add files via upload
1.Introduction 

The Indian Premier League (IPL) is a professional men's Twenty20 cricket league played in ten Indian cities by ten teams. The Board of Control for Cricket in India (BCCI) established the league in 2007. The teams are made up of players from all over the world. The IPL is the most-attended cricket league in the world, ranking sixth in average attendance among all sports leagues in 2014.The IPL became the world's first sporting event to be broadcast live on YouTube in 2010.

Data:
 The IPL dataset downloaded from Kaggle website (IPL Complete Dataset (2008-2020) | Kaggle). There are two csv files namely matches from 2008-2020 and Ball by ball statistics 2008-2020.The dataset includes match descriptions, results, winners, match players, a ball-by-ball dataset, and much more information about the IPL. The data dictionary below shows the key original variables along with new additional variables added to aid the analysis.
The dataset consists of two sheets.

Persona:
A company’s Betting Analyst wants to select the best value bets in the global cricket betting market, so he uses statistical expert analysis and cricket performance data to find value bets on behalf of clients. To find the best value bets, he is finding the answer to the following question.


Research Question:
1.	 Who are the top teams that can chasing the target even if their score at the 10th over is less than half of the target score? How many times have they done this? 
2.	 Does the winning toss and playing in the home venue has a relation to winning the game?


Requirements:
Here we specify the main requirements in terms of the relationships that must be visualized to answer each question. To answer the questions some insights are required, for that calculated fields are created, followed by brief design ideas for representations and interactions.
To answer Q1.the user needs to look at the tree map of biggest wins while chasing and the stacked bar chart showing the wins of chasing teams when the team didn’t make 50% of their target in the first 10 overs.
•	A Tree Map shows the biggest wins while chasing the score of their opponent.
•	A stacked bar chart showing the wins of chasing teams.

To answer Q2 the user needs to look at the stacked bar chart showing the relation between winning the toss at home venue and winning the match. There are 3 different stacked bar charts showing different relationships.

•	A stacked bar chart showing the relation between winning the toss at home venue and winning the match.

Initially, the prototype design answered the second question by showing a regression trend line, but during implementation, it was discovered that this was not the correct method to show correlation, even if doing one by one correlation between team and winning toss and correlation with home venue its visualization does not adequately justify the answer. Also, because it is a categorical variable, the chi square test may be an appropriate method to demonstrate that there is a relationship between winning the toss and playing at home in relation to winning the match.
demonstrating a positive relationship between winning the toss at home and winning the game The Chi-square test was used to verify this. The Chi-Square Test for Association is used to determine whether there is a relationship between two variables. It is, in fact, a hypothesis test for independence. The null hypothesis is that the two variables are not associated, i.e., independent. The alternate hypothesis is that the two variables are associated. The following sheets will show the process of chi-square test:
The p-value in statistical results is less than 0.05. We can reject the null hypothesis and conclude that there is a correlation between winning the toss at home venue and winning the game.